set-vit-tiny-det/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.3.0
albumentations>=1.3.0
pyyaml>=6.0
opencv-python>=4.5.0
matplotlib>=3.4.0
tqdm>=4.60.0
